Add requires_account_data question_routing_label
This adds a new requires_account_data question_routing_label enum in the answer model
and adds configuration to the question_routing_labels.yml file.

This enum replaces the personal_info tool for Claude answer generation, but is still
used in the OpenAI version of chat so we still require that enum value.