Retry Titan embedding generation
We've had an exception come through when generating Titan embeddings
where the input text exceeded the input token limit on Titan [1].

Titan supports either 50,000 characters OR 8,192 tokens for input [2]. We
already truncate to 50,000 characters, but there's no way for us to
know how many input tokens our text would convert to. So it's possible
that the input text is less than 50,000 characters but more than 8,192
tokens.

We used Tiktoken to calculate this for our OpenAI embeddings, but
there's no equivalent for other providers.

Currently an exception will just be raised if the token count is too
high, so this is an attempt to fix that. It will always truncate to
50,000 characters still, then it'll attempt an embedding on the
truncated text. If that fails, we'll try truncating to 80% of the
original text length. We'll continue truncating to 80% for 5 attempts.
If it's still too long then an exception is raised.

These numbers might need tweaking, but it feels like a good place to
start.

The vast majority of our chunks are within the limits, but sometimes a
bit chunk of HTML comes along (e.g. this content item [3]), which pushes it
over the limit.

[1]: https://govuk.sentry.io/issues/6726533431
[2]: https://docs.aws.amazon.com/bedrock/latest/userguide/titan-embedding-models.html
[3]: https://www.gov.uk/api/content/government/publications/balai-directive-approved-premises/list-of-bodies-institutes-and-centres-approved-for-intra-community-trade-in-animals