Stub GDS::SSO.test_user for mock auth
This changes the approach we use for mock authentication to work at the
GDS::SSO level rather than stubbing the controller. This change is
motivated by us moving authentication into the middleware for the API
and thus the stubs on the Controller are insufficient.

When we change the user we need to reset the session for web requests
with a session, as otherwise the existing logged in user will remain in
the session.

A consequence of this reset is that a previously dismissed cookie banner
will reappear. Therefore I have taken the step to moving this to be a
part of the system test following the logging in of a user and removed
as a tag that can be run before a test.