Use a signon uid rather than bearer token for rate limiting
This applies the new middleware to authenticate a user prior to
dispatching an API request so that we can use the uid of the user as the
discriminator for rate limiting.

As the UID value may be a semi-anonymous UUID or similar, I have
prefixed the discriminator with signon so it seems less of a random
value.

The motivation behind this change is:

- the bearer token is a secret, thus feels risky to log it or put it
  into memcache
- a single user could use multiple bearer tokens and thus get extra
  allowance
- having details of the user will allow us to log better the usage of a
  particular user.