Configure gds-sso for API
This changes the configuration so that any requests under the /api/
endpoint are treated as API calls and thus expect only bearer token
authentication.

As part of this change we now have different signon users used for
dev/test depending on whether it is session or bearer auth. Because of
this we now no longer need conversation-api as a permission on the basic
chat user. We do however need the dummy api user to have the
conversation-api permission and, for confusing reasons lost to history,
this is configured in a totally different way to the mock session user.