Fix GenerateTextEmbeddingJob when updating documents
Previously this job would call `index_document` which I'd assumed would
just update a single field if passed in.

However what it actually does is overwrite the entire document instead.

This adds a new `update_document` method in the ChunkedContentRepository
which will retain the original fields but set/update the attributes
passed in.