Return model in metrics for Claude answer composition steps
This updates the rephraser, question router and structured answer composition
steps to return the model used in the metrics.

This is useful for the admin UI to quickly see which model was used to
generate the LLM response for a given question.

I'll handle adding answer generation for OpenAI and the guardrails in
follow up commits.