Add E2E test for conversation API
We've had a few issues with the conversation API, around the link we return
for the answer_url. Our current tests check individual endpoints, and don't
check that the answer_url actaully works when used.

This adds an end-to-end test that checks the conversation API actually works
as expected when the requests are chained together.

I've chosen to go with a multi-request request spec rather than a system
spec. The reasoning for this is system specs really should be driven by
the user interface, which obviously an API doesn't have. We could have added
'type: :request' to a system spec, but the fact we would have had to do
that to get it working is a sign that it doesn't really fit the intended
use.

There's also prior art for a multi-request request spec in various gov.uk
code bases.