Add E2E tests for conversation API
We've had a few issues with the conversation API, around the link we return
for the answer_url. Our current tests check individual endpoints, and don't
check that the answer_url actually works when used.

This adds two an end-to-end tests:

1. mimics a conversation and hits the create and update convo endpoints
2. checks that pagination of a long conversation works as expected

I've chosen to go with a multi-request request spec rather than a system
spec. The reasoning for this is system specs really should be driven by
the user interface, which obviously an API doesn't have. We could have added
'type: :request' to a system spec, but the fact we would have had to do
that to get it working is a sign that it doesn't really fit the intended
use.

There's also prior art for a multi-request request spec in various gov.uk
code bases.

We have however, chosen to use given/when/then syntax for the tests since
it's more readable and better express the intent of the tests since they
cover multiple requests.

In order to follow the system spec conventions I've had to add some
exceptions to the file One that allows RSpec/NoExpectationExample
and another that allows instance variables to be used in the tests to
avoid lets and before blocks.