Use 'add_print_utility' to add print button data attributes
Previously, the only requirement for adding the print button data attributes to the header was that the conversation UI was active. This was causing JS errors, however, as the menu button is no longer being rendered and the JS was trying to find elements that didn't exist. Therefore, I added an 'add_print_utility' option to give a bit more control over when the data attributes are added and we can omit it when we don't require it