Add test to ensure question question routing labels match prompts config
We had an issue where we bumped the private gem with some breaking
changes, expecting CI to catch it, so we could apply changes there.

Instead it went green and got merged with the breaking changes. The reason
it didn't blow up is becase we weren't using the removed question routing
labels in our pipeline tests. So we don't try and dig for the prompt
in the config anywhere in our tests.

We have a test that checks the question routing labels enum values and
the question_routing_labels.yml config are in sync. I've added a similar
test to check that our prompt config is also in sync.

Unlike the previous test which checked for an exact match. I've just
checked that all the enum values are present in the prompt config keys
as we don't care if there are additional keys. For example, additional
keys might be added then used in a follow up PR post merge and we don't
want the tests to fail in that scenario.