Rename Total Prompt Tokens Method
This is a small change that renames the 'total_prompt_tokens' method within lib/bedrock_models.rb to 'claude_total_prompt_tokens'. This is due to the method being applicable only to Claude models, so renaming it to be indicative of this fact would assist in the readability of the codebase.