Add migration to add AnswerTopics table
This will be used to store the topics most closely related to an answer.
We will use an LLM tool call to determine the primary and optinal
secondary topic.

We will use this to set the primary and secondary topic on the model.
We will also store the metrics and LLM response on the model and
expose them in the admin interface.