Generate answer analysis topics after answer composition
This does all the required work to get topic tagging working for answers
generated by Claude.

It adds the AnswerAnalysisGeneration::TopicTagger which handles making
the LLM call and persiting AnswerAnalysis to the database based on the
response.

It also adds the AnswerTopicsJob which is responsible for calling the
TopicTagger after an answer has been composed provided the answer is
eligible for topic analysis and hasn't already been tagged with topics.

We will run this for all answer strategies, so we will need to make sure
that our OpenAI Heroku instance has the required env vars. I suspect it
will since we use titan for embeddings, but we will need to check.

It should be noted this is not the final implementation. Likely the tool
will change, and the system prompt will certaintly be updated. It's just
one that i've made myself.