Add AnswerInsights::TopicTagger
This does all the required work to get topic tagging working for answers
generated by Claude.

We will run this for all answer strategies, so we will need to make sure
that our OpenAI Heroku instance has the required env vars. I suspect it
will since we use titan for embeddings, but we will need to check.

It should be noted this is not the final implementation. Likely the tool
will change, and the system prompt will certaintly be updated. It's just
one that i've made myself.