Add AnswerTopic::Tagger and AnswerTopic::Claude::Tagger
This does all the required work to get topic tagging working for answers
generated by Claude.

It should be noted this is not the final implementation. Likely the tool
will change, and the system prompt will certaintly be updated. It's just
one that i've made myself.

It doesn't worry about exception handling or the OpenAI implementation.
These will be added in future commits.