Expose topics on admin question show page
This updates the question show page in the admin interface to
include the topics associated with the answer. It also add the topic LLM
response and metrics to the appropirate details components.

I've added a couple of helpers for the combined LLM responses and metrics.
to pull some of the logic out of the view.