Proof of concept for adding 2 sources to unanswerable answers
This is a proof of concept for adding the top 2 returned sources from
Opensearch to the message we return to users for unanswerable answers.

It is not intended to be productioh quality or be merged, but demonstrates
that it's pretty straight forward to do.

It would need a lot of TLC though because it needs to match the functionality
we use when rendering sources to users. For example using the title but
showing heading as well for duplicate titles.

To keep this simple i've simply rendered the exact url instead of the title.