Add answer_sources URL and Signon user email to BigQuery export
This is a temporary change to help with manual evaluation. We want to
be able to easily link users to the sources used and
be able to send users a list of their questions based on their email
address.

Alessia has requested that we link to producton URLs for sources used
so we've gsubbed links in the answe message to use them in addition to
using prod links in the answer source serialisation.