Updated dynamic canned answer for Claude llm_cannot_answer_message to return top 2 unique base_path as sources
Unique base_path are chosen because since we’re effectively signposting without full context (a bit in the dark), it's safer to point users to a broader page.
We are restricting the set of search results for which to identify the unique base_path to the first 5 search results because beyond that, we risk to return irrelevant pages with reputationally risky consequences - on top of being confusing for the users.