Implement homepage speech bubble icon
I opted for implementing 2 versions of the icon - one that animates and one that doesn't. By default, the animating svg will be shown and the static svg will be hidden. If the user activates the prefers-reduced-motion media query, the static svg will be shown and the animating svg will be hidden.