Ensure guardrails LLM response is the full response
When we ran the bigquery:export rake task, it failed because the answer
guardrails LLM response which we persist to the DB differs in type.

On successful calls, the response is a hash which contains various things
including the guardrail result, the LLM response, and token counts.
However, when the guardrail response is an error, we simply populate the
guardrails result e.g. "False | '1, 2'".

It's not clear to me why we did this initially, but the llm response as
as a whole will always be valid in this context as we've already had to
parse it to realise the returned string doesn't match our expected
pattern.

This updates the LLM response to just use the full response like we
do for all the other pipeline steps.

I've left in the guardrail result in the ResponseError so that it matches
the implementation of the Jailbreak Guardrails. Hopefully it also makes
it a little clearer to see the issue when an error is thrown.