Allow filtering of web created questions by signon user
I'm not really sure why this was added. I'm assuming it was before we
associated a Signon User with a conversation created via the web inferface.

I think it's probably because we expect there to be no user when/if we
go into Public Beta, but for now we since conversations can be only be
created with a Signon User we should allow filtering by that user in
both instances.