Ensure llm responses are stringified for BigQuery export
We ran into an issue where the BigQuery export was failing due to the format
of the guardrails LLM responses. Some were being persisted as hashes,
while others were strings. This inconsistency caused the export to fail.

To resolve this, we have updated the `serialize_for_export` method in the
`Answer` model to ensure that all LLM responses are stringified before
being exported.