Consider signon ID when rate limiting end user ID
When we rate limit the end user ID, we're just rate limiting whatever ID
the client gives us.

This opens up the possibility of a clash of IDs if/when we support
multiple clients.

To mitigate this, the key we use to throttle a request should include
the signon ID too. That way we'll get unique keys for users across
clients.

We want to first ensure that the header value is set; if it's not then
we don't do any throttling.