Add topic filters to questions index page
This allows filtering of questions based on primary and secondary topics.
I did consider returning from the secondary topic scope if it's the same
as the primary topic since they shouldn't really ever be the same, but
left it in in case we want to find a question that for some reason does
have the same topic for both.