Add topic filters to questions index page
This allows filtering of questions based on primary and secondary topics.
I did consider returning from the secondary topic scope if it's the same
as the primary topic since they shouldn't really ever be the same, but
left it in in case we want to find a question that for some reason does
have the same topic for both.

I've added a test to ensure that we can load the topics correctly from
the private repo configuration. We don't want it to fail silently if
changes are made to the structure of the configuration in the private
repo.