Add middleware to normalise path in Rack
When load testing GOV.UK Chat we noticed that we were getting uncaught
exceptions. These were occurring because an unusual path had been used for the
load test //api/conversation/e4abee9c-1f11-4dec-9cb5-c20ed4b55f92.

What was interesting about this was that this request was raising an
error due to failing to match the Api::AuthMiddleware [1] but
managing to match the Rack::Attack middleware [2]. It transpired that
these different pieces of middleware were using a different value for
path as Rack::Attack normalises the path [3] before passing it on to
other middleware.

As this inconsistency risks subtle bugs, and because Rails itself does a
normalising process at the point of routing a request, I've added
additional middleware to normalise a path before we make any middleware
decisions based on the path. Warden::Manager is the first piece of
middleware that makes use of it as it uses to be determine whether we
have an API request or not.

[1]: https://github.com/alphagov/govuk-chat/blob/0a81713ce586a3f8e69fbbcd940125d33cccb7f6/lib/api/auth_middleware.rb#L9-L11
[2]: https://github.com/alphagov/govuk-chat/blob/0a81713ce586a3f8e69fbbcd940125d33cccb7f6/config/initializers/rack_attack.rb#L11-L15
[3]: https://github.com/rack/rack-attack/blob/467770882daa6f3865cc207c8b5dfdbc4028d7cb/lib/rack/attack.rb#L108