Exclude error and jailbreak guardrails statuses from topic tagging
We only want to carry out analysis on answers that have successfully
completed the answer composition pipeline and had legitimate questions.

There are several reasons for this, but the two most important are:

1. It will become increasingly complex in implementation and cognitive load
   to have some error statuses that are tagged and some aren't. It's far
   easier to just exclude all error statuses from topic tagging.
2. Based on the ~700 questions we had for manual eval, answers with
   error statuses made up less than 0.5% of the total. This means that
   excluding them from topic tagging will not have a significant impact on
   the overall topic tagging process.

We've also chosen to exclude Jailbreak guardrails from topic tagging,
because we've catgegorised the user intent as malicous in some way and
therefore don't want to include them in any topic analysis we do.