Add request specs for API middleware
We've not got a growing collection of API middleware that is
individually unit tested but previously lacked integration tests.

This test file is to try capture those tests - which gives me a place to
put a regression test for the unusual path issue we experienced.

It's somewhat tricky, particularly with mock GDS SSO, to actually test
differences in authentication between bearer and non-bearer tokens. I
thought the easiest way was to assert it based on a failed request as
I wasn't sure what to assert on for a successful request.