Include rate limiting in API middleware request tests
This adds tests for the individual aspects of rate limiting that are
applied to routes as a whole.

We can likely move some of our individual route rack attack tests into
here to avoid as many running at a time and instead have just a single
check that they return rate limit headers.