Send rate limit usage to Prometheus for API user
This extends the current rate limiting middleware to log rate limit
usage for an API user.

We want some way of knowing how close API users are to their request
quota, and doing this via Prometheus/Grafana seems like a good way to
go.

So in the middleware we inspect the rate limit data we've got from
Rack::Attack. If the rate limit data is related to the throttling for
the API user's read methods then we log the percentage of read requests
remaining for that user in Prometheus. Same for write requests.

We could log the raw number, i.e. the number of requests remaining, but
this would be a bit meaningless when looking at the chart in Grafana
unless you know the actual rate limit figures. It feels like logging the
percentage remaining means we can create a chart that's easily
understandable at a glance.

Using a percentage as the metric also means that if we change the rate
limits, we just need to change one number in this app and we won't need
to make any changes to the chart.