Add healthcheck for Bedrock
Adds a healthcheck for Bedrock and adds it to the existing healthchecks.

We want to make a lightweight request in this healthcheck (i.e. no
calling of actual models). To do that, we're including the
"aws-sdk-bedrock" gem so that we can make a request to list the models
available to us.

The healthcheck verifies that all the models we're using in the
application - as defined in the `BedrockModels` module - are available
to us within Bedrock.

This makes it much easier to verify that we've enabled the right set of
models in Bedrock. When we want to use a new model, we need to enable it
in the AWS Console, and it can be easy to forget to do this in all
environments. This way we'll be notified automatically.