Add healthcheck for Bedrock
Adds a healthcheck for Bedrock and adds it to the existing healthchecks.

It makes a single request to the `invoke_model` endpoint using the
Titan embedding model. This is a model with a short latency (rather than
using Claude Sonnet, for example), so should return a response quickly.

We just verify that an exception isn't raised to determine whether
Bedrock is healthy or not.