Use service navigation component in admin pages
In the changelog for the govuk_publishing_components gem there is a breaking
change. [See this PR](https://github.com/alphagov/govuk_publishing_components/pull/4934).

The layout_header menu has been replaced with a service navigation component.
This means that we need to import the service navigation component in our
admin stylesheet and JS files, and update our tests to use the new
components class names.