Add postcode PII check to the PII validator
We've been asked to add a check for UK postcodes to the PII validator.
This adds a fairly simple regex to match on UK postcode formats.

It doesn't include a full list of actual postcodes and only checks for
a very specific format (i.e.1 or no spaces between the first and second
part of the postcode).