Run onboarding animation only once
If a user has already completed onboarding, it may cause frustration to have to go through it again when the user refreshes the page or clears their chat. Therefore, I've added a cookie to detect when the onboarding animation has completed and will prevent it from running again as long as the cookie exists.