Support v0 of API
A previous commit bumped the API version from v0 to v1 and moved all the
API routes to `/api/v1` instead of `/api/v0`.

We still need to support v0 for a little while though, in order for the
app team to upgrade to v1 at their own pace.

This commit reinstates the `/api/v0` routes but points them to the new
v1 controllers. Rather than doing any fancy metaprogramming or loops in
the route file, it's just a copy and paste job. The idea being that that
once we need to remove support for v0, we can just delete this block of
routes.

A couple of other things needed doing to support v0 routes:

* The conversations controller returns URLs in some of its responses. If
  the request is for `/api/v0/*` then we should return `/api/v0/*` URLs
  in the response. So we do a simple regex on the path to determine
  which URL to return. Again, this is intentionally verbose so it's
  easier to clean up
* We need the Committee middleware to support both `/api/v0` and
  `/api/v1` prefixes.

The v1 request specs have been fully copy/pasted here with a couple of
minor changes (to test that `/api/v0/*` URLs are returned for example).
Again this will aid in the cleanup when we stop supporting v0; we can
just delete this entire directory of specs.