Add temporary "non_llm_answer" strategy
This strategy (which is configured as the default for this branch) will
generate an answer without making any LLM calls.

It's for the SREs to be able to do load testing without incurring costs
from any LLM providers.

The `sleep` calls emulate the latency that we see in a normal
LLM-based strategy.