Add e2e system spec to track user interactions in admin area
We have tests that cover a users interactions with chat e2e and we have
tests that check when conversation data is present it is displayed correctly
in the admin area. But we don't have one that checks the end-to-end flow
of user interactions from the chat interface to the admin area.

This adds a test that:
- logs you in as admin with the web app permission
- asks a question
- gets an answer
- gives feedback on the answer
- visits the admin area
- navigates to the question show page
- checks the answer, feedback and topics are displayed