Move shared .exportable functionality into shared example
We have the concept on .exportable on various models that we send to
BigQuery. These largely have similar impelementations. They grab records
between two datetimes. This functionality can be extracted into a shared
example which we can reuse.

The exception to this is the 'Question' model, which has additional logic
for:

- preloading the conversation, answer and sources assoicated with it which
we want to test
- it uses the answers created_at instead of the questions created_at

After thinking for a while about whether i should use the shared example
for questions due to these differences I've chosen to use the shared example
for shared functionality. I've left the tests that test additional
functionality in the questions .exportable describe block.

I'm not wed to this though as it's a bit harder to parse what's actually
going on than leaving the tests as is. So if anyone disagrees i'm happy
to go with that.