Add config for sidekiq answer and default processes
This adds config for separate Sidekiq processes for answer and default jobs.
I've chosen to implement these in separate configuration files rather
than combining them and using env vars to set the queue names for
simplicity.

This way we can simply use the separate configuration files for each process
and not have to worry setting extra en vars in helm.

I've left the env vars for configuring concurrency in place. It looks
like we're not using these env vars in helm at the moment, but i'm assuming
it's been left in place for a reason. I can see we also have the option
to configure the db pool. I can remove them if we decide we don't need them
anymore.

The plan is to leave the current sidekiq config file in place until it's
not being used anymore in helm.