API: Require end user ID header
This header value is currently optional. If it's set then it'll
associate a conversation with the user ID specified. If it's not, then
the conversation is just associated with the signon user.

One DPIA requirement for Chat is that a user can request deletion of
their data. Without requiring this header value we wouldn't be able to
find the conversations for a given user. So we need to require it.