Update topic tagger to return result object
We want to be able to run auto eval on the topics we tag questions with.
Currently, the topic tagger persists the answer analysis results in the database.

This is undesirable because we don't want to persist records to the db
during auto eval.

This updates the topic tagger to take a user_question and returns a
results object with the primary & secondary topics, as well as the LLM
response and metrics.

The job now handles persisting the analysis to the database based on
the results object.

It also handles passing the question message or rephrased question to
the topic tagger.