Reduce threshold of alerting on rate limit usage
On review, we felt that configuring a slack notification to occur at 90%
of rate limit usage wouldn't be an early enough indicator of any trends
that indicate a likelihood to hit a rate limit. This has instead been
lowered to 75%.

If this is something that is actually triggered we may want to consider
moving it from this codebase to be triggered by monitoring Prometheus as
this is liable to quite noisy should we receive a few requests that go
over a limit.