Add completeness filter for questions in admin ui
Now we're populating the completeness field on answers during answer
generation, we can use this to filter questions in the admin UI based on
the completeness of the answer.

It should be noted that questions generated using OpenAI won't ever show
when this filter has a value as they will always have a nil completeness.