Daily Slack notification for chat activity
When the next trial goes live, we want to post a daily message to Slack
which shows an overview of the questions asked the previous day. This
gives everyone visibility of how the app is being used.

This extends the existing SlackPoster class to add a method which posts
the previous day's Chat activity.

The building of the message is done in a separate class as it's a little
complex, and was a bit much for a single method in the SlackPoster
class.

It finds answers created any time during the previous day which have a
certain status and builds a mesage to post to Slack with links to the
right place in the admin panel for those answers.