Add initial table of answer_source_chunks
The purpose of this table is to store the data from opensearch at the
time of us generating an answer. The reason to store this is so that we
can re-access this data as part of any auto-evaluation so we can compare
what was used with the result.

This has been put into a separate table so that the data can be reused
each time the same piece of content is being used. It is intended that
once this table is fully introduced then data from the answer source
table will be removed.

The null false values for this are based on my understanding of which
fields we could expect to be null from the Publishing API - I'm
potentially introducing a risk of a crash here if I've assumed a field
can't be null when it actually can be in our search index (as that has
no assurances) so we will have to monitor it.