Add columns to store search and weighted scores
I'm adding extra data here that hasn't been asked for but I'm including
it for completeness. Now that we're planning to store which chunks are
used in an answer it seems logical we may also want to know what score
those items were given in searching.

I've named the generic field we have of score to search score here to
distinguish it from the weighted score. I couldn't find a previously
used qualifier.

I chose a float data type as that seemed closer to the original data
format OpenSearch uses than a decimal - but both are probably fine.