Add helper methods for finding and creating answer source chunks
We expect answer_source_chunk models to only be created if an existing
one doesn't exist to avoid unnecessary data storage. To achieve this
I've added a find_or_create_from_search_result method that can create an
item and be resilient to one being created at the same time.

Using find_or_create_by! is resilient to there being a collision on the
unique index when create_or_find_by! attempts [1] so this should allow
for concurrency to not produce problems, however this requires this code
to be ran outside a transaction and be atomic - which is how we
currently create an answer.

[1]: https://github.com/rails/rails/blob/bb429032aa0daf7bad2a8e6ac17ec8c5934e6d19/activerecord/lib/active_record/relation.rb#L293-L308