Create AnswerSourceChunk objects when building sources
This changes the approach to building sources for an Answer to utilise
data from AnswerSourceChunk. This is the first step towards removing the
storing of chunk data from an AnswerSource record.

This does something that could be quite confusing, it writes a
AnswerSourceChunk to the database straight away when one doesn't exist.
This means there could be a situation where an Answer source is not
persisted to database and the chunk is.

The motivation for doing this is to avoid us storing chunks multiple
times and avoid risking exceptions by delaying the saving. Using this
approach means that as soon as an Answer wants to write the usage of a
chunk one is created and if any concurrent Answers are being generated
they will look up this newly created one.

This could create problems were we to wrap the AnswerComposition process
in a transaction, as non-unique collisions would not be spotted until
the transaction is committed. If this was the case we'd probably have to
consider a different approach where we tolerate record collisions.

Following this change and migration of existing sources we can remove
the storage of extra data from an AnswerSource.