Migration to backfill AnswerSourceChunk data
This migration copies data from AnswerSource records to
AnswerSourceChunk to enable migrating from using the former for search
result data to the latter. I've done this as a migration rather than a
rake task as it is needed to be run on any systems with

This is quite a slow migration if the dataset is large so there is a
chance it could crash in some environments. But as it is not in a
transaction it should be able to resume and complete on a retry.