Resolve null / empty array confusion with heading_hierarchy
This changes the column definition to default to an empty array and
disallow. The motivation for this is to avoid having to consider whether
it can be both nil and an empty array. Since it seems to be impossible
for it not to have the possibility of being an empty array it simplifies
things if it can't also be nil.

There is a chance this could break if any databases already have a null
value for this column but from my understanding of the code this isn't
currently possible so I don't expect this to fail.