Add topics graph to metrics page
We recently started tagging primary and an optional secondary topic to
answers. We've exposed these on the question show page and by filters on
the questions index page. It would be useful to see a graph of the topics
tagged to answers over time on the metrics page too.

This adds that graph using the same pattern as the other graphs on the
metrics page.

I've updated this to use the answers.created_at rather than
answer_analyses.created_at based on feedback from code review.

One thing i ran into that was a bit of an issue is that i've had to use
2.times {} rather than using a create_list to set created_at on the answer
to avoid the unique constraint.

I could use something like this:
```
create_list(:answer_analysis, 2) do |analysis|
  analysis.answer.created_at = 2.days.ago
  analysis.primary_topic = "tax"
end
```

etc. but it feels easier to just use 2.times { create stuff }.