Change absolute URL to absolute path
This resolves an issue flagged on our ITHC where this particular form
could be subject to a DNS rebinding attack because of it's use of an
absolute URL. Using an absolute path resolves this issue.

We may still want to investigate having an allow list of hosts to
prevent the risk of such attacks further.